refactor: extract InvoicesFeature component and simplify InvoicesPage
