+++ app/users/dashboard/finance/invoices/page.tsx
@@ -1,529 +1,17 @@
 'use client';
 
-import { useState, useMemo } from 'react';
 import { mockInvoices, mockProjects } from '@/components/shared/mock-data';
-import { AppLayout, Pagination, SearchAndFilter } from '@/components/common';
-import { Button } from '@/components/ui/button';
-import { Badge } from '@/components/ui/badge';
-import { Checkbox } from '@/components/ui/checkbox';
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-} from '@/components/ui/card';
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from '@/components/ui/table';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
-import {
-  FileText,
-  DollarSign,
-  Calendar,
-  CheckCircle,
-  Clock,
-} from 'lucide-react';
-import Link from 'next/link';
-import { format } from 'date-fns';
-import {
-  InvoiceType,
-  InvoiceStatus,
-  invoiceTypeLabels,
-  invoiceStatusLabels,
-} from '@/types/finance/invoice';
+import { InvoicesFeature } from '@/features/invoices';
 
-const getStatusColor = (status: InvoiceStatus) => {
-  switch (status) {
-    case InvoiceStatus.paid: {
-      return 'bg-green-100 text-green-800 dark:bg-green-900/20 dark:text-green-400';
-    }
-    case InvoiceStatus.partiallyPaid: {
-      return 'bg-blue-100 text-blue-800 dark:bg-blue-900/20 dark:text-blue-400';
-    }
-    case InvoiceStatus.pending: {
-      return 'bg-yellow-100 text-yellow-800 dark:bg-yellow-900/20 dark:text-yellow-400';
-    }
-    case InvoiceStatus.sent: {
-      return 'bg-purple-100 text-purple-800 dark:bg-purple-900/20 dark:text-purple-400';
-    }
-    case InvoiceStatus.draft: {
-      return 'bg-zinc-100 text-zinc-800 dark:bg-zinc-800 dark:text-zinc-400';
-    }
-    case InvoiceStatus.overdue: {
-      return 'bg-red-100 text-red-800 dark:bg-red-900/20 dark:text-red-400';
-    }
-    case InvoiceStatus.cancelled: {
-      return 'bg-zinc-100 text-zinc-800 dark:bg-zinc-800 dark:text-zinc-400';
-    }
-    case InvoiceStatus.disputed: {
-      return 'bg-orange-100 text-orange-800 dark:bg-orange-900/20 dark:text-orange-400';
-    }
-    default: {
-      return 'bg-zinc-100 text-zinc-800 dark:bg-zinc-800 dark:text-zinc-400';
-    }
-  }
-};
-
-const getTypeColor = (type: InvoiceType) => {
-  switch (type) {
-    case InvoiceType.purchase: {
-      return 'bg-blue-100 text-blue-800 dark:bg-blue-900/20 dark:text-blue-400';
-    }
-    case InvoiceType.sales: {
-      return 'bg-green-100 text-green-800 dark:bg-green-900/20 dark:text-green-400';
-    }
-    case InvoiceType.expense: {
-      return 'bg-orange-100 text-orange-800 dark:bg-orange-900/20 dark:text-orange-400';
-    }
-    case InvoiceType.service: {
-      return 'bg-indigo-100 text-indigo-800 dark:bg-indigo-900/20 dark:text-indigo-400';
-    }
-    default: {
-      return 'bg-zinc-100 text-zinc-800 dark:bg-zinc-800 dark:text-zinc-400';
-    }
-  }
-};
-
+/**
+ * Invoices page - thin routing layer
+ * Fetches/provides data and renders the feature component
+ */
 export default function InvoicesPage() {
-  const [searchQuery, setSearchQuery] = useState('');
-  const [statusFilter, setStatusFilter] = useState<string>('all');
-  const [typeFilter, setTypeFilter] = useState<string>('all');
-  const [currentPage, setCurrentPage] = useState(1);
-  const [itemsPerPage, setItemsPerPage] = useState(10);
-  const [selectedIds, setSelectedIds] = useState<number[]>([]);
+  // In a real application, this would fetch data from an API
+  // For now, we're using mock data
+  const invoices = mockInvoices;
+  const projects = mockProjects;
 
-  // Filter invoices based on search and filters
-  const filteredInvoices = useMemo(() => {
-    return mockInvoices.filter((invoice) => {
-      // Search filter
-      const searchLower = searchQuery.toLowerCase();
-      const matchesSearch =
-        !searchQuery ||
-        invoice.invoiceNumber.toLowerCase().includes(searchLower) ||
-        invoice.notes?.toLowerCase().includes(searchLower) ||
-        invoice.paymentTerms?.toLowerCase().includes(searchLower);
-
-      // Status filter
-      const matchesStatus =
-        statusFilter === 'all' || invoice.status === statusFilter;
-
-      // Type filter
-      const matchesType = typeFilter === 'all' || invoice.type === typeFilter;
-
-      return matchesSearch && matchesStatus && matchesType;
-    });
-  }, [searchQuery, statusFilter, typeFilter]);
-
-  // Pagination
-  const startIndex = (currentPage - 1) * itemsPerPage;
-  const endIndex = startIndex + itemsPerPage;
-  const paginatedInvoices = filteredInvoices.slice(startIndex, endIndex);
-  const totalPages = Math.ceil(filteredInvoices.length / itemsPerPage);
-
-  // Calculate stats
-  const totalInvoices = mockInvoices.length;
-  const paidInvoices = mockInvoices.filter(
-    (i) => i.status === InvoiceStatus.paid
-  ).length;
-  const pendingInvoices = mockInvoices.filter(
-    (i) => i.status === InvoiceStatus.pending || i.status === InvoiceStatus.sent
-  ).length;
-  const totalAmount = mockInvoices.reduce((sum, i) => sum + i.totalAmount, 0);
-
-  // Selection handlers
-  const handleSelectAll = (checked: boolean) => {
-    if (checked) {
-      setSelectedIds(paginatedInvoices.map((i) => i.id));
-    } else {
-      setSelectedIds([]);
-    }
-  };
-
-  const handleSelectOne = (id: number, checked: boolean) => {
-    if (checked) {
-      setSelectedIds([...selectedIds, id]);
-    } else {
-      setSelectedIds(selectedIds.filter((selectedId) => selectedId !== id));
-    }
-  };
-
-  const isAllSelected =
-    paginatedInvoices.length > 0 &&
-    selectedIds.length === paginatedInvoices.length;
-
-  const hasActiveFilters = Boolean(
-    searchQuery || statusFilter !== 'all' || typeFilter !== 'all'
-  );
-
-  const clearFilters = () => {
-    setSearchQuery('');
-    setStatusFilter('all');
-    setTypeFilter('all');
-    setCurrentPage(1);
-  };
-
-  return (
-    <AppLayout>
-      <div className="space-y-4 sm:space-y-6">
-        {/* Header */}
-        <div className="mb-8 flex items-start justify-between">
-          <div>
-            <h1 className="mb-2 text-3xl font-bold text-zinc-900 dark:text-zinc-100">
-              Invoices
-            </h1>
-            <p className="text-zinc-600 dark:text-zinc-400">
-              Manage invoices, track payments, and monitor outstanding balances
-            </p>
-          </div>
-          <Button asChild>
-            <Link href="/users/dashboard/finance/invoices/new">
-              <FileText className="mr-2 h-4 w-4" />
-              New Invoice
-            </Link>
-          </Button>
-        </div>
-
-        {/* Statistics Cards */}
-        <div className="mb-8 grid grid-cols-1 gap-6 md:grid-cols-4">
-          <Card>
-            <CardHeader className="pb-3">
-              <CardDescription>Total Invoices</CardDescription>
-            </CardHeader>
-            <CardContent>
-              <div className="flex items-center justify-between">
-                <div className="flex h-12 w-12 items-center justify-center rounded-lg bg-blue-100 dark:bg-blue-900/20">
-                  <FileText className="h-6 w-6 text-blue-600 dark:text-blue-400" />
-                </div>
-                <span className="text-2xl font-bold text-zinc-900 dark:text-zinc-100">
-                  {totalInvoices}
-                </span>
-              </div>
-            </CardContent>
-          </Card>
-
-          <Card>
-            <CardHeader className="pb-3">
-              <CardDescription>Paid Invoices</CardDescription>
-            </CardHeader>
-            <CardContent>
-              <div className="flex items-center justify-between">
-                <div className="flex h-12 w-12 items-center justify-center rounded-lg bg-green-100 dark:bg-green-900/20">
-                  <CheckCircle className="h-6 w-6 text-green-600 dark:text-green-400" />
-                </div>
-                <span className="text-2xl font-bold text-zinc-900 dark:text-zinc-100">
-                  {paidInvoices}
-                </span>
-              </div>
-            </CardContent>
-          </Card>
-
-          <Card>
-            <CardHeader className="pb-3">
-              <CardDescription>Pending/Sent</CardDescription>
-            </CardHeader>
-            <CardContent>
-              <div className="flex items-center justify-between">
-                <div className="flex h-12 w-12 items-center justify-center rounded-lg bg-yellow-100 dark:bg-yellow-900/20">
-                  <Clock className="h-6 w-6 text-yellow-600 dark:text-yellow-400" />
-                </div>
-                <span className="text-2xl font-bold text-zinc-900 dark:text-zinc-100">
-                  {pendingInvoices}
-                </span>
-              </div>
-            </CardContent>
-          </Card>
-
-          <Card>
-            <CardHeader className="pb-3">
-              <CardDescription>Total Value</CardDescription>
-            </CardHeader>
-            <CardContent>
-              <div className="flex items-center justify-between">
-                <div className="flex h-12 w-12 items-center justify-center rounded-lg bg-purple-100 dark:bg-purple-900/20">
-                  <DollarSign className="h-6 w-6 text-purple-600 dark:text-purple-400" />
-                </div>
-                <span className="text-2xl font-bold text-zinc-900 dark:text-zinc-100">
-                  ₹{(totalAmount / 1_000_000).toFixed(1)}M
-                </span>
-              </div>
-            </CardContent>
-          </Card>
-        </div>
-
-        {/* Search and Filters */}
-        <SearchAndFilter
-          searchValue={searchQuery}
-          onSearchChange={(value) => {
-            setSearchQuery(value);
-            setCurrentPage(1);
-          }}
-          searchPlaceholder="Search by invoice number, notes..."
-          hasActiveFilters={hasActiveFilters}
-          onClearFilters={clearFilters}
-          filters={[
-            {
-              placeholder: 'Status',
-              options: [
-                { value: 'all', label: 'All Statuses' },
-                {
-                  value: InvoiceStatus.draft,
-                  label: invoiceStatusLabels[InvoiceStatus.draft],
-                },
-                {
-                  value: InvoiceStatus.pending,
-                  label: invoiceStatusLabels[InvoiceStatus.pending],
-                },
-                {
-                  value: InvoiceStatus.sent,
-                  label: invoiceStatusLabels[InvoiceStatus.sent],
-                },
-                {
-                  value: InvoiceStatus.partiallyPaid,
-                  label: invoiceStatusLabels[InvoiceStatus.partiallyPaid],
-                },
-                {
-                  value: InvoiceStatus.paid,
-                  label: invoiceStatusLabels[InvoiceStatus.paid],
-                },
-                {
-                  value: InvoiceStatus.overdue,
-                  label: invoiceStatusLabels[InvoiceStatus.overdue],
-                },
-                {
-                  value: InvoiceStatus.cancelled,
-                  label: invoiceStatusLabels[InvoiceStatus.cancelled],
-                },
-                {
-                  value: InvoiceStatus.disputed,
-                  label: invoiceStatusLabels[InvoiceStatus.disputed],
-                },
-              ],
-              value: statusFilter,
-              onChange: (value) => {
-                setStatusFilter(value);
-                setCurrentPage(1);
-              },
-            },
-            {
-              placeholder: 'Type',
-              options: [
-                { value: 'all', label: 'All Types' },
-                {
-                  value: InvoiceType.purchase,
-                  label: invoiceTypeLabels[InvoiceType.purchase],
-                },
-                {
-                  value: InvoiceType.sales,
-                  label: invoiceTypeLabels[InvoiceType.sales],
-                },
-                {
-                  value: InvoiceType.expense,
-                  label: invoiceTypeLabels[InvoiceType.expense],
-                },
-                {
-                  value: InvoiceType.service,
-                  label: invoiceTypeLabels[InvoiceType.service],
-                },
-              ],
-              value: typeFilter,
-              onChange: (value) => {
-                setTypeFilter(value);
-                setCurrentPage(1);
-              },
-            },
-          ]}
-        />
-
-        {/* Results Summary */}
-        <div className="mb-4 flex items-center justify-between">
-          <p className="text-sm text-zinc-600 dark:text-zinc-400">
-            Showing {startIndex + 1} to{' '}
-            {Math.min(endIndex, filteredInvoices.length)} of{' '}
-            {filteredInvoices.length} invoices
-          </p>
-          <div className="flex items-center space-x-2">
-            <span className="text-sm text-zinc-600 dark:text-zinc-400">
-              Rows per page:
-            </span>
-            <Select
-              value={itemsPerPage.toString()}
-              onValueChange={(value) => {
-                setItemsPerPage(Number(value));
-                setCurrentPage(1);
-              }}
-            >
-              <SelectTrigger className="w-[70px]">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="5">5</SelectItem>
-                <SelectItem value="10">10</SelectItem>
-                <SelectItem value="20">20</SelectItem>
-                <SelectItem value="50">50</SelectItem>
-                <SelectItem value="100">100</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-        </div>
-
-        {/* Invoices Table */}
-        {filteredInvoices.length > 0 ? (
-          <Card>
-            <CardContent className="p-0">
-              <Table>
-                <TableHeader>
-                  <TableRow>
-                    <TableHead className="w-12">
-                      <Checkbox
-                        checked={isAllSelected}
-                        onCheckedChange={handleSelectAll}
-                        aria-label="Select all"
-                      />
-                    </TableHead>
-                    <TableHead>Invoice Number</TableHead>
-                    <TableHead>Type</TableHead>
-                    <TableHead>Project</TableHead>
-                    <TableHead>Amount</TableHead>
-                    <TableHead>Issue Date</TableHead>
-                    <TableHead>Due Date</TableHead>
-                    <TableHead>Status</TableHead>
-                    <TableHead>Balance</TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {paginatedInvoices.map((invoice) => {
-                    return (
-                      <TableRow
-                        key={invoice.id}
-                        className="hover:bg-muted/50 cursor-pointer"
-                        onClick={() =>
-                          (globalThis.location.href = `/dashboard/finance/invoices/${invoice.id}`)
-                        }
-                      >
-                        <TableCell onClick={(e) => e.stopPropagation()}>
-                          <Checkbox
-                            checked={selectedIds.includes(invoice.id)}
-                            onCheckedChange={(checked) =>
-                              handleSelectOne(invoice.id, checked as boolean)
-                            }
-                            aria-label={`Select ${invoice.invoiceNumber}`}
-                          />
-                        </TableCell>
-                        <TableCell>
-                          <div className="flex items-center space-x-3">
-                            <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-linear-to-br from-blue-500 to-blue-600">
-                              <FileText className="h-5 w-5 text-white" />
-                            </div>
-                            <div>
-                              <p className="font-medium text-zinc-900 dark:text-zinc-100">
-                                {invoice.invoiceNumber}
-                              </p>
-                              {invoice.paymentTerms && (
-                                <p className="text-xs text-zinc-500 dark:text-zinc-500">
-                                  Terms: {invoice.paymentTerms}
-                                </p>
-                              )}
-                            </div>
-                          </div>
-                        </TableCell>
-                        <TableCell>
-                          <Badge className={getTypeColor(invoice.type)}>
-                            {invoiceTypeLabels[invoice.type]}
-                          </Badge>
-                        </TableCell>
-                        <TableCell>
-                          <span className="text-sm font-medium text-zinc-700 dark:text-zinc-300">
-                            {mockProjects.find(
-                              (p) => p.id === invoice.projectId
-                            )?.projectName || 'Unknown Project'}
-                          </span>
-                        </TableCell>
-                        <TableCell>
-                          <span className="font-semibold text-zinc-900 dark:text-zinc-100">
-                            ₹{invoice.totalAmount.toLocaleString('en-IN')}
-                          </span>
-                        </TableCell>
-                        <TableCell>
-                          <div className="flex items-center space-x-2 text-sm text-zinc-600 dark:text-zinc-400">
-                            <Calendar className="h-3 w-3 text-zinc-400" />
-                            <span>
-                              {format(invoice.issueDate, 'dd MMM yyyy')}
-                            </span>
-                          </div>
-                        </TableCell>
-                        <TableCell>
-                          <span className="text-sm text-zinc-700 dark:text-zinc-300">
-                            {format(invoice.dueDate, 'dd MMM yyyy')}
-                          </span>
-                        </TableCell>
-                        <TableCell>
-                          <Badge className={getStatusColor(invoice.status)}>
-                            {invoiceStatusLabels[invoice.status]}
-                          </Badge>
-                        </TableCell>
-                        <TableCell>
-                          <span
-                            className={`font-semibold ${
-                              invoice.balanceAmount > 0
-                                ? 'text-orange-600 dark:text-orange-400'
-                                : 'text-green-600 dark:text-green-400'
-                            }`}
-                          >
-                            ₹{invoice.balanceAmount.toLocaleString('en-IN')}
-                          </span>
-                        </TableCell>
-                      </TableRow>
-                    );
-                  })}
-                </TableBody>
-              </Table>
-            </CardContent>
-
-            {/* Pagination Controls */}
-            <Pagination
-              currentPage={currentPage}
-              totalPages={totalPages}
-              onPageChange={setCurrentPage}
-            />
-          </Card>
-        ) : (
-          <Card>
-            <CardContent className="flex h-64 items-center justify-center">
-              <div className="text-center">
-                <FileText className="mx-auto h-12 w-12 text-zinc-400" />
-                <h3 className="mt-4 text-lg font-semibold text-zinc-900 dark:text-zinc-100">
-                  No invoices found
-                </h3>
-                <p className="mt-2 text-sm text-zinc-600 dark:text-zinc-400">
-                  {hasActiveFilters
-                    ? 'Try adjusting your filters'
-                    : 'Get started by creating a new invoice'}
-                </p>
-                {hasActiveFilters && (
-                  <Button
-                    variant="outline"
-                    className="mt-4"
-                    onClick={clearFilters}
-                  >
-                    Clear Filters
-                  </Button>
-                )}
-              </div>
-            </CardContent>
-          </Card>
-        )}
-      </div>
-    </AppLayout>
-  );
+  return <InvoicesFeature invoices={invoices} projects={projects} />;
 }
