+++ features/invoices/index.ts
@@ -1,0 +1,1 @@
+export { InvoicesFeature } from './components/invoices-feature';
